pyproject.toml
[project]
name = "quant-data-open"
version = "0.1.0"
description = "Quant investment data collection and visualization"
authors = [
    {name = "Junho", email = "pinedance@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Data collection
    "finance-datareader>=0.9.50",
    "yfinance>=0.2.59",
    "requests>=2.31.0",

    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Web scraping
    "beautifulsoup4>=4.12.0",  # bs4 package
    "lxml>=4.9.0",  # HTML parser for BeautifulSoup

    # Templating & config
    "jinja2>=3.1.2",
    "pyyaml>=6.0.1",

    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",

    # Visualization (optional, for future use)
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "autopep8>=2.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["core"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
